% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-corpus.R
\name{create_ttt_corpus}
\alias{create_ttt_corpus}
\title{ttt_create_corpus}
\usage{
create_ttt_corpus(data_dir = "./data")
}
\arguments{
\item{data_dir}{Directory where data are to be stored (will be created if it
does not exist).}
}
\value{
A \pkg{quanteda} corpus representing the text contents of all `pdf`
documents contained in `data_dir`.
}
\description{
Create a corpus by reading a directory of pre-downloaded `pdf` files.
}
