% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{dl_canada}
\alias{dl_canada}
\title{dl_canada}
\usage{
dl_canada(data_dir = "./data", quiet = FALSE)
}
\arguments{
\item{data_dir}{Directory where data are to be stored (will be created if it
does not exist).}

\item{quiet}{If `TRUE`, display progress information on screen}
}
\value{
`TRUE` if successful
}
\description{
Download annual Indian Affairs Reports for Canada
}
