% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot.keyness_annual}
\alias{plot.keyness_annual}
\title{plot.keyness_annual}
\usage{
\method{plot}{keyness_annual}(x, ...)
}
\arguments{
\item{x}{A 'keyness_annual' object}

\item{...}{Ingored here}
}
\value{
Nothing (generates plot)
}
\description{
plot.keyness_annual
}
