% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{ttt_fit_topics}
\alias{ttt_fit_topics}
\title{ttt_topics}
\usage{
ttt_fit_topics(tok, years = NULL, ntopics = 20, topic = NULL,
  filename = NULL, quiet = FALSE)
}
\arguments{
\item{tok}{A \pkg{quanteda} `tokens` object}

\item{years}{If specified, restrict topic models to specified years only
(requires `tok` to include a `docvars` of "year" or similar).}

\item{ntopics}{Number of topics to be fitted. This should be reasonably
large, to allow words not strongly associated with any primary topics to be
allocated to latter, effectively meaningless topics.}

\item{topic}{If specified, construct topic models around the specified
phrase.}

\item{filename}{If specified, the topic model is saved to the nominated file
and can be re-loaded with `x <- readRDS(filename)`.}

\item{quiet}{If `TRUE`, display progress information on screen.}
}
\value{
An \pkg{topicmodels} object of class `LDA`.
}
\description{
Fit sentence-based topic models for all of some of a corpus, with the ability
to fit topics around a specified word or phrase.
}
\note{
This function may take a long time to execute; please be patient.
}
\examples{
library(quanteda)
dat <- data_corpus_inaugural \%>\% # from quanteda
     corpus_reshape (to = "sentences") # convert documents to sentences
tok <- tokens (dat, remove_numbers = TRUE, remove_punct = TRUE,
               remove_separators = TRUE)
tok <- tokens_remove(tok, stopwords("english"))
x <- ttt_fit_topics (tok, ntopics = 5)
topicmodels::get_terms(x, 20)
x <- ttt_fit_topics (tok, years = 1789:1900, ntopics = 5)
}
