% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness.R
\name{ttt_keyness}
\alias{ttt_keyness}
\title{ttt_keyness}
\usage{
ttt_keyness(x, word = "school", window = 10)
}
\arguments{
\item{x}{A \pkg{quanteda} `tokens` list, obtained by `tokens(corpus)`.}

\item{word}{The key word for which associations with other words are to be
calculated.}

\item{window}{Passed to \pkg{quanteda} `tokens_keep/remove` functions; the
number of words surrounding each instance of `word` to be considered in
measures of assocation.}
}
\value{
A \pkg{quanteda} `keyness` object listing words (`features`) and
associated keyness statistics.
}
\description{
Keyness metrics for associations with specified key words; see
\url{https://en.wikipedia.org/wiki/Keyword_(linguistics)}
}
