% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness.R
\name{ttt_keyness2}
\alias{ttt_keyness2}
\title{ttt_keyness2}
\usage{
ttt_keyness2(x, word)
}
\arguments{
\item{x}{Either a single \pkg{quanteda} `keyness` object, as for example
returned from \link{ttt_keyness}, or a list of such objects, as for example
return from \link{ttt_keyness_annual}.}

\item{word}{Secondary word for which the association with the keyword used in
\link{ttt_keyness} is to be extracted.}
}
\value{
A \pkg{quanteda} `keyness` object filtered to the specified
associations only (see Note).
}
\description{
Extract association with a specified word from the results of
\link{ttt_keyness}. The first function calculates associations with a
specified word; this function reduces those only to associations with a
second specified word.
}
\note{
For single `keyness` objects, this function is merely a very thin
wrapped around \pkg{dplyr} `filter`. For annual lists of `keyness` objects
returned from \link{ttt_keyness_annual}, each year is filtered to the
specified associations only, and the list converted to a single `keyness`
`data.frame` with an additional `year` column.
}
\examples{
# prepare a corpus of quanteda tokens:
dat <- quanteda::data_corpus_inaugural
tok <- quanteda::tokens (dat, remove_numbers = TRUE, remove_punct = TRUE,
               remove_separators = TRUE)
tok <- quanteda::tokens_remove(tok, quanteda::stopwords("english"))
# then use that to extract keyword associations:
x <- ttt_keyness (tok, "politic*")
# and filter to specified association only
x <- ttt_keyness2 (x, "petty")
}
