% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness.R
\name{ttt_keyness_annual}
\alias{ttt_keyness_annual}
\title{ttt_keyness_annual}
\usage{
ttt_keyness_annual(x, word = "school", window = 10,
  remove_keyword = FALSE)
}
\arguments{
\item{x}{A \pkg{quanteda} `tokens` list, obtained by `tokens(corpus)`.}

\item{word}{The key word for which associations with other words are to be
calculated.}

\item{window}{Passed to \pkg{quanteda} `tokens_keep/remove` functions; the
number of words surrounding each instance of `word` to be considered in
measures of assocation.}

\item{remove_keyword}{If `TRUE`, remove the specified keyword from results,
leaving only associations with that word not the word itself.}
}
\value{
A list of \pkg{quanteda} `keyness` objects listing words (`features`)
and associated keyness statistics; one list item per year (where able to be
calcualted).
}
\description{
Keyness metrics for associations with specified key words, evaluated on an
annual basis; see
\url{https://en.wikipedia.org/wiki/Keyword_(linguistics)}
}
\note{
Only those years for which `x` contains the nominated `word` will
return entries, and thus the return length of this function may be less than
the number of years in the corpus.
}
\examples{
# prepare a corpus of quanteda tokens:
dat <- quanteda::data_corpus_inaugural
tok <- quanteda::tokens (dat, remove_numbers = TRUE, remove_punct = TRUE,
               remove_separators = TRUE)
tok <- quanteda::tokens_remove(tok, quanteda::stopwords("english"))
# then use that to extract keyword associations:
\dontrun{
x <- ttt_keyness_annual (tok, "school")
x <- ttt_keyness_annual (tok, "school", remove_keyword = TRUE)
}
}
