% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{ttt_fit_topics}
\alias{ttt_fit_topics}
\title{ttt_topics}
\usage{
ttt_fit_topics(dat, years = NULL, ntopics = 20, topic = NULL,
  filename = NULL)
}
\arguments{
\item{dat}{A temporally-structued \pkg{quanteda} corpus with a `docar` of
`year`}

\item{years}{If specified, restrict topic models to specified years only}

\item{ntopics}{Number of topics to be fitted. This should be reasonably
large, to allow words not strongly associated with any primary topics to be
allocated to latter, effectively meaningless topics.}

\item{topic}{If specified, construct topic models around the specified
phrase.}

\item{filename}{If specified, the topic model is saved to the nominated file
and can be re-loaded with `x <- readRDS(filename)`.}
}
\value{
An \pkg{topicmodels} object of class `LDA`.
}
\description{
Fit sentence-based topic models for all of some of a corpus, with the ability
to fit topics around a specified word or phrase.
}
\note{
This function may take a long time to execute; please be patient.
}
