% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness.R
\name{ttt_keyness}
\alias{ttt_keyness}
\title{ttt_keyness}
\usage{
ttt_keyness(x, word = "school", window = 10, remove_keyword = FALSE)
}
\arguments{
\item{x}{A \pkg{quanteda} `tokens` list, obtained by `tokens(corpus)`.}

\item{word}{The key word for which associations with other words are to be
calculated.}

\item{window}{Passed to \pkg{quanteda} `tokens_keep/remove` functions; the
number of words surrounding each instance of `word` to be considered in
measures of assocation.}

\item{remove_keyword}{If `TRUE`, remove the specified keyword from results,
leaving only associations with that word not the word itself.}
}
\value{
A \pkg{quanteda} `keyness` object listing words (`features`) and
associated keyness statistics.
}
\description{
Keyness metrics for associations with specified key words; see
\url{https://en.wikipedia.org/wiki/Keyword_(linguistics)}
}
\examples{
# prepare a corpus of quanteda tokens:
dat <- quanteda::data_corpus_inaugural
tok <- quanteda::tokens (dat, remove_numbers = TRUE, remove_punct = TRUE,
               remove_separators = TRUE)
tok <- quanteda::tokens_remove(tok, quanteda::stopwords("english"))
# then use that to extract keyword associations:
x <- ttt_keyness (tok, "school")
head (x, n = 20)
x <- ttt_keyness (tok, "politic*")
head (x, n = 20)
}
